+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/FileScanSourceOpExec.scala
@@ -1,15 +1,12 @@
-package edu.uci.ics.texera.workflow.operators.source.scan.text
+package edu.uci.ics.texera.workflow.operators.source.scan
 
 import edu.uci.ics.texera.workflow.common.operators.source.SourceOperatorExecutor
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils
-import org.apache.commons.compress.archivers.{ArchiveEntry, ArchiveStreamFactory}
 import org.apache.commons.io.IOUtils.toByteArray
 
 import java.io._
-import java.nio.file.{Files, Paths}
 import java.util.zip.ZipFile
-import scala.collection.mutable
 import scala.jdk.CollectionConverters.{asScalaIteratorConverter, enumerationAsScalaIteratorConverter}
 
 class FileScanSourceOpExec private[text] (val desc: FileScanSourceOpDesc)
@@ -17,23 +14,29 @@ class FileScanSourceOpExec private[text] (val desc: FileScanSourceOpDesc)
 
   @throws[IOException]
   override def produceTexeraTuple(): Iterator[Tuple] = {
-    val filePath = desc.filePath.getOrElse(throw new RuntimeException("Empty file path accepted."))
-
     val fileEntries: Iterator[InputStream] = if (desc.unzip) {
-      val zipReader = new ZipFile(filePath)
+      val zipReader = new ZipFile(desc.filePath.get)
       zipReader.entries()
         .asScala
         .filterNot(entry => entry.getName.startsWith("__MACOSX"))
         .map(entry => zipReader.getInputStream(entry))
     } else {
-      Iterator(new FileInputStream(filePath))
+      Iterator(new FileInputStream(desc.filePath.get))
     }
-
 
     if (desc.attributeType.isSingle) {
       fileEntries.map(entry => produceSingleTuple(entry))
     } else {
-      fileEntries.flatMap(entry => produceMultiTuples(entry))
+      fileEntries.flatMap(entry => new BufferedReader(new InputStreamReader(entry, desc.encoding.getCharset))
+        .lines()
+        .iterator()
+        .asScala
+        .slice(
+          desc.fileScanOffset.getOrElse(0),
+          desc.fileScanOffset.getOrElse(0) + desc.fileScanLimit.getOrElse(Int.MaxValue)
+        )
+        .map(line => produceSingleTuple(new ByteArrayInputStream(line.getBytes))
+        ))
     }
   }
 
@@ -49,19 +52,6 @@ class FileScanSourceOpExec private[text] (val desc: FileScanSourceOpDesc)
         }
       )
       .build()
-  }
-
-  private def produceMultiTuples(stream: InputStream): Iterator[Tuple] = {
-    new BufferedReader(new InputStreamReader(stream, desc.encoding.getCharset))
-      .lines()
-      .iterator()
-      .asScala
-      .slice(
-        desc.fileScanOffset.getOrElse(0),
-        desc.fileScanOffset.getOrElse(0) + desc.fileScanLimit.getOrElse(Int.MaxValue)
-      )
-      .map(line => produceSingleTuple( new ByteArrayInputStream(line.getBytes))
-      )
   }
 
   override def open(): Unit = {}
