+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/source/scan/text/TextSourceOpDesc.scala
@@ -2,12 +2,9 @@ package edu.uci.ics.texera.workflow.operators.source.scan.text
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
-import com.kjetland.jackson.jsonSchema.annotations.{
-  JsonSchemaInject,
-  JsonSchemaString,
-  JsonSchemaTitle
-}
+import com.kjetland.jackson.jsonSchema.annotations.{JsonSchemaInject, JsonSchemaString, JsonSchemaTitle}
 import edu.uci.ics.texera.workflow.common.metadata.annotations.HideAnnotation
+import edu.uci.ics.texera.workflow.operators.source.scan.FileAttributeType
 
 /**
   * TextSourceOpDesc is a trait holding commonly used properties and functions
