+++ core/amber/src/test/scala/edu/uci/ics/texera/workflow/operators/source/scan/text/FileScanSourceOpDescSpec.scala
@@ -2,7 +2,7 @@ package edu.uci.ics.texera.workflow.operators.source.scan.text
 
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.{AttributeType, Schema}
-import edu.uci.ics.texera.workflow.operators.source.scan.FileDecodingMethod
+import edu.uci.ics.texera.workflow.operators.source.scan.{FileAttributeType, FileDecodingMethod, FileScanSourceOpDesc, FileScanSourceOpExec}
 import org.scalatest.BeforeAndAfter
 import org.scalatest.flatspec.AnyFlatSpec
